Create SAScodefor2007match.txt
This file, if saved as a .sas program, is designed to match the March 2008 CPS ASEC and the 2007 IRS SOI PUF data following the methodology used by John O'Hare for 2004. The code has been altered to use variables names, rather than positions in the dataset, in many cases, and now includes a wider variety of CPS data, rather than just the 50 variables that were included in the previous process. This was used internally at the Heritage Foundation for some modeling work and may be useful for the OSPC model as well.